11/15 commit
Accomplished goal and stretch goal. Created a model for an inning of
baseball, accounting for transition probabilities.  Currently the only
statistic produced is amount of runs scored. For now, model only uses
team averages.

Algorithm currently does not account for the possibility of a double
play, will add in a later iteration. Also, team statistics are hard
coded in right now, will add functionality for reading from a csv file
in a later iteration.